Refactor: rename data.local to data.room and improve Kover exclusions
1. Renamed package: memos.data.local → memos.data.room
   - Makes it clear this is Room database code
   - Follows same pattern as .platform.* for platform-specific code

2. Updated Kover exclusions to use universal patterns:
   - Added `*.room.*` exclusion (like `*.platform.*`)
   - Removed hardcoded package name exclusion
   - Now any future Room code will be automatically excluded

3. Coverage: 94.5% (1792/1896 lines)
   - Removed brittle per-class exclusions
   - System is now more maintainable and scalable
